+++ src/components/day/calendarAllDayEvent.component.ts
@@ -10,7 +10,10 @@ import { CalendarEvent } from 'calendar-utils';
 @Component({
   selector: 'mwl-calendar-all-day-event',
   template: `
-    <ng-template #defaultTemplate>
+    <ng-template
+      #defaultTemplate
+      let-event="event"
+      let-eventClicked="eventClicked">
       <div
         class="cal-all-day-event"
         [style.backgroundColor]="event.color.secondary"
@@ -18,6 +21,7 @@ import { CalendarEvent } from 'calendar-utils';
         <mwl-calendar-event-actions [event]="event"></mwl-calendar-event-actions>
         <mwl-calendar-event-title
           [event]="event"
+          [customTemplate]="eventTitleTemplate"
           view="day"
           (mwlClick)="eventClicked.emit()">
         </mwl-calendar-event-title>
@@ -36,6 +40,8 @@ export class CalendarAllDayEventComponent {
   @Input() event: CalendarEvent;
 
   @Input() customTemplate: TemplateRef<any>;
+
+  @Input() eventTitleTemplate: TemplateRef<any>;
 
   @Output() eventClicked: EventEmitter<any> = new EventEmitter();
 }