+++ src/components/week/calendarWeekViewEvent.component.ts
@@ -10,7 +10,13 @@ import { WeekViewEvent } from 'calendar-utils';
 @Component({
   selector: 'mwl-calendar-week-view-event',
   template: `
-    <ng-template #defaultTemplate>
+    <ng-template
+      #defaultTemplate
+      let-weekEvent="weekEvent"
+      let-tooltipPlacement="tooltipPlacement"
+      let-eventClicked="eventClicked"
+      let-tooltipTemplate="tooltipTemplate"
+      let-tooltipAppendToBody="tooltipAppendToBody">
       <div
         class="cal-event"
         [style.backgroundColor]="weekEvent.event.color.secondary"
@@ -22,6 +28,7 @@ import { WeekViewEvent } from 'calendar-utils';
         <mwl-calendar-event-actions [event]="weekEvent.event"></mwl-calendar-event-actions>
         <mwl-calendar-event-title
           [event]="weekEvent.event"
+          [customTemplate]="eventTitleTemplate"
           view="week"
           (mwlClick)="eventClicked.emit()">
         </mwl-calendar-event-title>
@@ -47,6 +54,8 @@ export class CalendarWeekViewEventComponent {
   @Input() tooltipAppendToBody: boolean;
 
   @Input() customTemplate: TemplateRef<any>;
+
+  @Input() eventTitleTemplate: TemplateRef<any>;
 
   @Input() tooltipTemplate: TemplateRef<any>;